Revert HIDLs in the vendor repo to OMB27.43-68
In the update OMB27.43-70, a new requirement of symbol " _ZN7android8hardware7details14gServiceSidMapE " was introduced. None of the 8.0 ROMs sources were updated with this symbol as it was abandoned.
Some 8.1 and 9.0 ROMs were updated with this symbol, but not all. So, for now we are reverting the conflicting blobs.